Implement change syntax Reorganise source tree
