+++ src/widgets/mod.rs
@@ -4,6 +4,7 @@ mod palette_list;
 pub mod editor_view;
 pub mod bottom_panel;
 pub mod window;
+pub mod text_buffer;
 
 pub use empty::*;
 pub use extension::*;
