+++ src/widgets/text_buffer/mod.rs
@@ -4,6 +4,7 @@ mod edit_stack;
 mod file;
 pub mod position;
 pub mod rope_utils;
+pub mod syntax;
 
 pub use edit_stack::*;
