+++ src/widgets/text_buffer/rope_utils.rs
