+++ src/widgets/text_buffer/syntax.rs
